==> experiments/01-scibert/03-train-scibert.ipynb <==
# SCIBERT

trained on Zora

import spacy
import pandas as pd
from pathlib import Path

from spacy.cli.train import train as spacy_train

train_set = 'zora'
test_set = 'zora'
model_name = 'scibert'

base_path = Path.cwd().parent

# Edit paths as needed:
config_path = base_path / f"cfg/spacy/{model_name}.cfg"
model_out_path = base_path / f"data/models/ft-{model_name}-{train_set}"
train_path, dev_path, test_path = (
    base_path / f"data/spacy_docs/{train_set}_train.spacy",
    base_path / f"data/spacy_docs/{train_set}_dev.spacy",
    base_path / f"data/spacy_docs/{test_set}_test.spacy"
)

# train
spacy_train(
    config_path,
    output_path=model_out_path,
    use_gpu=0,
    overrides={
        "paths.train": str(train_path),
        "paths.dev": str(dev_path)
    }
)

# test
best_model = base_path / f"data/models/ft-scibert-zora/model-best"
test_cats = [str(i) for i in range(1, 18)]
output_file= f"{best_model}/test_{test_set}_eval.json"

!python3 -m spacy benchmark accuracy --gpu-id 0 \
  {best_model}/ {test_path} \
  --output {output_file}


def get_test_df(file):
  test_df = pd.read_csv(file, sep='\t', keep_default_na=False,
                        index_col=0, encoding='utf-8')
  test_df = test_df.astype({'sdg': 'string',
                            'abstract': 'string'})
  test_df.drop(columns=['faculty', 'year'], inplace=True)

  return test_df

# Edit `test_df` path:
test_df = get_test_df(f"data/train_test/{test_set}_test_clean.tsv")

X_test = test_df['abstract'].values
y_test = test_df['sdg'].values

nlp = spacy.load(best_model)
print("Making predictions....")

spacy_probs = [doc.cats for doc in nlp.pipe(X_test)]
print("Done making predictions!")

# For each item, select the label to which the model has assigned the highest probability:
preds = []
probs = []
for label_probs_dict in spacy_probs:
    pred, prob = max(label_probs_dict.items(), key=lambda x: x[1])
    preds.append(pred)
    probs.append(prob)

preds = pd.Series(preds)
probs = pd.Series(probs)

preds_df = pd.DataFrame({'abstract': X_test,
                         'label': y_test,
                         'prediction': preds,
                         'probability': probs})
# Align original indices
preds_df.index = test_df.index

preds_df = preds_df.astype({'abstract': 'string',
                            'label': 'int',
                            'prediction': 'int',
                            'probability': 'float'})

preds_file = base_path / f"data/{train_set}-{test_set}_preds.tsv"
preds_df.to_csv(preds_file, sep='\t', encoding='utf-8')
